Se crea el muro Principal  para se actualizado
