+++ config/apollo.js
@@ -4,8 +4,8 @@ import fetch from 'node-fetch';
 import { setContext } from 'apollo-link-context';
 
 const httpLink = createUploadLink({
-	//uri: 'http://localhost:4000/graphql',
-	uri: 'https://splayserver.herokuapp.com/',
+	uri: 'http://localhost:4000/graphql',
+	//uri: 'https://splayserver.herokuapp.com/',
 	fetch
 });
