+++ pages/wall.jsx
@@ -1,16 +1,15 @@
-import styled from 'styled-components';
-
+import Home from '../components/organims/Home';
 import HeaderLogged from '../components/organims/Navbar/HeaderLogged';
 
-export default function wall() {
+export default function index(props) {
+	const { auth } = props;
+
 	return (
 		<React.Fragment>
 			<HeaderLogged />
-			<MensajeMuroStyle>Bienvenido al Muro de Splay pronto en Produccion feliz dia...</MensajeMuroStyle>
+			<Home auth={auth} />
 		</React.Fragment>
 	);
 }
 
 //css
-
-const MensajeMuroStyle = styled.h1`margin-top: 140px;`;
